demo/requirements.txt
Flask
tensorflow
numpy
opencv-python
Pillow